+++ backend/api/src/main/kotlin/io/tolgee/hateoas/notification/NotificationSettingGroupModel.kt
@@ -1,0 +1,9 @@
+package io.tolgee.hateoas.notification
+
+import org.springframework.hateoas.RepresentationModel
+import java.io.Serializable
+
+data class NotificationSettingGroupModel(
+  var inApp: Boolean,
+  var email: Boolean,
+) : RepresentationModel<NotificationSettingGroupModel>(), Serializable