+++ backend/data/src/main/kotlin/io/tolgee/development/testDataBuilder/data/AdministrationTestData.kt
@@ -1,20 +1,23 @@
 package io.tolgee.development.testDataBuilder.data
 
 import io.tolgee.development.testDataBuilder.builders.TestDataBuilder
+import io.tolgee.development.testDataBuilder.builders.UserAccountBuilder
 import io.tolgee.model.UserAccount
 
 class AdministrationTestData {
   lateinit var admin: UserAccount
   lateinit var user: UserAccount
+  var adminBuilder: UserAccountBuilder
 
   val root =
     TestDataBuilder().apply {
-      addUserAccount {
-        username = "admin@admin.com"
-        name = "Peter Administrator"
-        role = UserAccount.Role.ADMIN
-        admin = this
-      }
+      adminBuilder =
+        addUserAccount {
+          username = "admin@admin.com"
+          name = "Peter Administrator"
+          role = UserAccount.Role.ADMIN
+          admin = this
+        }
 
       addUserAccount {
         username = "user@user.com"