+++ e2e/cypress/common/shared.ts
@@ -13,8 +13,9 @@ export const allScopes: Scope[] = [
 ];
 
 export const clickAdd = () => {
-  cy.wait(100);
-  cy.xpath(getAnyContainingAriaLabelAttribute('add')).click();
+  cy.xpath(getAnyContainingAriaLabelAttribute('add'))
+    .should('be.visible')
+    .click();
 };
 
 export const getPopover = () => {
@@ -128,6 +129,25 @@ export const toggleInMultiselect = (
   });
   dismissMenu();
   waitForGlobalLoading();
+};
+
+export const assertMultiselect = (chainable: Chainable, values: string[]) => {
+  chainable.find('div').first().click();
+
+  getPopover().within(() => {
+    getPopover()
+      .get('li.MuiMenuItem-root')
+      .each(($li) => {
+        const labelText = $li.find('.MuiListItemText-primary').text();
+        const input = cy.wrap($li).find('input');
+        if (values.includes(labelText)) {
+          input.should('be.checked');
+        } else {
+          input.should('not.be.checked');
+        }
+      });
+  });
+  dismissMenu();
 };
 
 export const getInputByName = (name: string): Chainable => {